pyproject.toml
[tool.poetry]
name = "larin-les27"
version = "0.1.0"
description = ""
authors = ["MrLerich <93153582+MrLerich@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "larin_les27"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.3"
psycopg2 = "^2.9.5"
pillow = "^9.3.0"
djangorestframework = "^3.14.0"

[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
